Added the new controller for area and broker.
implemented the accessToken and refreshToken using httpOnly cookie during signin functionality and stored in cookies,
which used to mitigate the CSRF attacks and invalidate this accessToken and refreshToken during logout functionality.

accessToken is used for access the api and when accessToken is expired,then frontend interceptor will send refreshToken,
which used to validate the refreshToken and send new accessToken,at some point of time,even refresh token is also expired at time,
directly in fronend it will logout,so user will login and get new accessToken and refreshToken.

implemented the public and private certificates for validating the accessToken and refreshToken.

implemented the chokidar and config dependencies for config refresh and implemented the db.config.js and auth.config.js,but
testing is done for db.config.js,when I update the host for db.config.js it is connecting to new host automatically.

during this db config update,existing require will delete with cache require and get reload the model with new sequelize object and will used,
in all the services.

pass the accessToken for all endPoint and verification of accessToken is also happening for all routes and endpoints.

Modify the accountService to support the in-memory and persistant cache,still need to do field mapping service.

changed the event.js to location.controller.js and location.routes.js and evently this services will deleted once dashboard fucntionality is ready.

create generateSecrets.js,which is used to generate random secret,which is used in forget password functionality.